+++ packages/ui/components/calendar-switch/index.ts
@@ -1,1 +1,1 @@
-export { CalendarSwitchComponent } from "./CalendarSwitch";
+export { CalendarSwitchComponent, type ICalendarSwitchProps } from "./CalendarSwitch";
